breaking: rename TreeLoader:open to :addSpace
