+++ lib/loaders/TreeLoader/addPlaceholder.js
@@ -1,7 +1,7 @@
 module.exports = function (id, content) {
-  // @TreeLoader:placeholder(id, content)
+  // @TreeLoader:addPlaceholder(id, content)
   //
-  // Open a placeholder space. Placeholders do not trigger
+  // Add a placeholder space. Placeholders do not trigger
   // recursive loading.
   //
   // Parameters:
@@ -11,5 +11,5 @@ module.exports = function (id, content) {
   //     a Component
   //
 
-  return this.open(id, content, true)
+  return this.addSpace(id, content, true)
 }
