+++ lib/loaders/TreeLoader/addSpace.js
