+++ lib/loaders/TreeLoader/index.js
@@ -83,11 +83,11 @@ const TreeLoader = function (config) {
   //       return response.json()
   //     })
   //     .then(data => {
-  //       loader.open(id, data.html)
+  //       loader.addSpace(id, data.html)
   //       callback(null)
   //     })
   //     .catch(error => {
-  //       loader.open(id, error.message)
+  //       loader.addSpace(id, error.message)
   //       console.error('Error:', error)
   //       callback(error)
   //     })
@@ -157,7 +157,12 @@ const TreeLoader = function (config) {
   // ```
   //
   // Now you have the TreeLoader constructed.
-  // You still need to build a driver for it. See examples.
+  // You still need to build a driver for it.
+  // The driver is a function ran at each viewport idle event
+  // or similar, non-realtime schedule.
+  // Its purpose is to find our current location in the tree
+  // and open/close the TreeLoader nodes accordingly.
+  // See examples.
   //
 
   if (!config.viewport || !config.viewport.isViewport) {
@@ -238,16 +243,16 @@ proto.isTreeLoader = true
 // Inherit
 emitter(proto)
 
-proto.init = require('./init')
+proto.addPlaceholder = require('./addPlaceholder')
+proto.addSpace = require('./addSpace')
 proto.closeChild = require('./closeChild')
 proto.closeChildren = require('./closeChildren')
 proto.closeNeighbors = require('./closeNeighbors')
 proto.closeParent = require('./closeParent')
-proto.open = require('./open')
+proto.init = require('./init')
 proto.openChild = require('./openChild')
 proto.openChildren = require('./openChildren')
 proto.openNeighbors = require('./openNeighbors')
 proto.openParent = require('./openParent')
-proto.placeholder = require('./placeholder')
 proto.remapChildren = require('./remapChildren')
 proto.removeSpace = require('./removeSpace')
